Bugfix: Doesn't delete random messages

+++ mod/antibot.js
@@ -1,10 +1,17 @@
 module.exports = async (client) => {
   const typings = {};
+  let allowBypass = true;
+  let bypassUsed = [];
+  setTimeout(() => {
+    allowBypass = false;
+    bypassUsed = [];
+  }, 60000);
   client.on("typingStart", async (typing) => {
     if (!typings[typing.user.id]) typings[typing.user.id] = [];
     typings[typing.user.id].push({
       channel: typing.channel.id,
       startedAt: typing.startedAt,
+      startedTimestamp: typing.startedTimestamp
     });
   });
   client.on("messageCreate", async (msg) => {
@@ -13,18 +20,11 @@ module.exports = async (client) => {
       !msg.member.bot &&
       (await client.enableAntibot.get(msg.guild.id))
     ) {
-      let typing = typings[msg.member.id]
-        ? typings[msg.member.id].filter((t) => t.channel == msg.channel.id)[0]
-        : null;
-      typings[msg.member.id] = typings[msg.member.id]
-        ? typings[msg.member.id].filter((t) => t.channel != msg.channel.id)
-        : [];
-      if (
-        !typing ||
-        typing.startedAt > new Date(Date.now() - msg.content.length / 10) ||
-        msg.embeds
-      )
-        return msg.delete();
+      if (msg.embeds) return msg.delete();
+      if (!typings[msg.member.id] && (!allowBypass || bypassUsed.indexOf(msg.author.id) < 0)) return msg.delete();
+      if (!typings[msg.member.id].some((t) => t.channel == msg.channel.id) && (!allowBypass || bypassUsed.indexOf(msg.member.id) < 0)) return msg.delete();
+      if (allowBypass) bypassUsed.push(msg.member.id);
+      typings[msg.member.id] = typings[msg.member.id].filter((t) => t.channel != msg.channel.id);
     }
   });
 };
